Support setup.py-less python packages.
