+++ disperse/__main__.py
@@ -17,7 +17,6 @@
 
 from datetime import datetime
 from glob import glob
-import json
 import logging
 import os
 import re
@@ -25,7 +24,6 @@ import subprocess
 import sys
 import time
 from typing import Optional, List, Tuple
-from urllib.request import urlopen, Request
 from urllib.parse import urlparse
 
 from github import Github  # type: ignore
@@ -42,11 +40,19 @@ from breezy.revision import NULL_REVISION
 from silver_platter.workspace import Workspace
 
 
-from . import NoUnreleasedChanges, version_string
+from . import NoUnreleasedChanges
 from .news_file import (
     NewsFile,
     news_find_pending,
     )
+from .python import (
+    pypi_discover_urls,
+    UploadCommandFailed,
+    DistCommandFailed,
+    upload_python_artifacts,
+    create_python_artifacts,
+    create_setup_py_artifacts,
+)
 
 
 DEFAULT_CI_TIMEOUT = 7200
@@ -65,20 +71,6 @@ class RecentCommits(Exception):
 
 
 class VerifyCommandFailed(Exception):
-
-    def __init__(self, command, retcode):
-        self.command = command
-        self.retcode = retcode
-
-
-class UploadCommandFailed(Exception):
-
-    def __init__(self, command, retcode):
-        self.command = command
-        self.retcode = retcode
-
-
-class DistCommandFailed(Exception):
 
     def __init__(self, command, retcode):
         self.command = command
@@ -262,69 +254,6 @@ def check_new_revisions(
             raise NoUnreleasedChanges()
 
 
-def upload_python_artifacts(local_tree, pypi_paths):
-    command = [
-        "twine", "upload", "--non-interactive",
-        "--sign"] + pypi_paths
-    try:
-        subprocess.check_call(command, cwd=local_tree.abspath("."))
-    except subprocess.CalledProcessError as e:
-        raise UploadCommandFailed(command, e.returncode)
-
-
-def create_python_artifacts(local_tree):
-    # Import setuptools, just in case it tries to replace distutils
-    import setuptools  # noqa: F401
-    from distutils.core import run_setup
-
-    orig_dir = os.getcwd()
-    try:
-        os.chdir(local_tree.abspath('.'))
-        result = run_setup(
-            local_tree.abspath("setup.py"), stop_after="config")
-    finally:
-        os.chdir(orig_dir)
-    pypi_paths = []
-    is_pure = (
-        not result.has_c_libraries()  # type: ignore
-        and not result.has_ext_modules())  # type: ignore
-    if is_pure:
-        try:
-            subprocess.check_call(
-                ["./setup.py", "bdist_wheel"],
-                cwd=local_tree.abspath(".")
-            )
-        except subprocess.CalledProcessError as e:
-            raise DistCommandFailed(
-                "setup.py bdist_wheel", e.returncode)
-        wheels_glob = 'dist/%s-%s-*-any.whl' % (
-            result.get_name().replace('-', '_'),  # type: ignore
-            result.get_version())  # type: ignore
-        wheels = glob(
-            os.path.join(local_tree.abspath('.'), wheels_glob))
-        if not wheels:
-            raise AssertionError(
-                'setup.py bdist_wheel did not produce expected files. '
-                'glob: %r, files: %r' % (
-                    wheels_glob,
-                    os.listdir(local_tree.abspath('dist'))))
-        pypi_paths.extend(wheels)
-    else:
-        logging.warning(
-            'python module is not pure; not uploading binary wheels')
-    try:
-        subprocess.check_call(
-            ["./setup.py", "sdist"], cwd=local_tree.abspath(".")
-        )
-    except subprocess.CalledProcessError as e:
-        raise DistCommandFailed("setup.py sdist", e.returncode)
-    sdist_path = os.path.join(
-        "dist", "%s-%s.tar.gz" % (
-            result.get_name(), result.get_version()))  # type: ignore
-    pypi_paths.append(sdist_path)
-    return pypi_paths
-
-
 def release_project(   # noqa: C901
         repo_url: str, *, force: bool = False,
         new_version: Optional[str] = None,
@@ -486,6 +415,8 @@ def release_project(   # noqa: C901
             ws.local_tree.branch.tags.set_tag(
                 tag_name, ws.local_tree.last_revision())
         if ws.local_tree.has_filename("setup.py"):
+            pypi_paths = create_setup_py_artifacts(ws.local_tree)
+        elif ws.local_tree.has_filename("pyproject.toml"):
             pypi_paths = create_python_artifacts(ws.local_tree)
         else:
             pypi_paths = None
@@ -634,32 +565,6 @@ def create_github_release(repo, tag_name, version, description):
     repo.create_git_release(
         tag=tag_name, name=version, draft=False, prerelease=False,
         message=description or (f'Release {version}.'))
-
-
-def pypi_discover_urls(pypi_user):
-    import xmlrpc.client
-    client = xmlrpc.client.ServerProxy('https://pypi.org/pypi')
-    ret = []
-    for relation, package in client.user_packages(pypi_user):
-        req = Request(
-            f'https://pypi.org/pypi/{package}/json',
-            headers={'Content-Type': f'disperse/{version_string}'})
-        with urlopen(req) as f:
-            data = json.load(f)
-        project_urls = data['info']['project_urls']
-        if project_urls is None:
-            logging.warning(f'Project {package} does not have project URLs')
-            continue
-        for key, url in project_urls.items():
-            if key == 'Repository':
-                ret.append(url)
-                break
-            parsed_url = urlparse(url)
-            if (parsed_url.hostname == 'github.com' and
-                    parsed_url.path.strip('/').count('/') == 1):
-                ret.append(url)
-                break
-    return ret
 
 
 def validate_config(path):
