+++ setup.cfg
@@ -23,6 +23,7 @@ install_requires =
     protobuf>=2.4
     pygithub
     silver-platter>=0.5.2
+    build
 
 [options.entry_points]
 console_scripts =
